+++ src/main/java/com/bookjob/member/dto/request/UpdateNicknameRequest.java
@@ -1,4 +1,4 @@
-package com.bookjob.member.dto;
+package com.bookjob.member.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
