+++ src/main/java/com/bookjob/member/dto/response/MyPageResponse.java
@@ -1,4 +1,4 @@
-package com.bookjob.member.dto;
+package com.bookjob.member.dto.response;
 
 public record MyPageResponse (
         String nickname,
