+++ src/main/java/com/bookjob/member/dto/response/MyPostingsInBoardResponse.java
@@ -1,4 +1,6 @@
-package com.bookjob.member.dto;
+package com.bookjob.member.dto.response;
+
+import com.bookjob.member.dto.MyPostingsInBoard;
 
 import java.util.List;
