+++ src/main/java/com/bookjob/member/service/MemberReadService.java
@@ -3,8 +3,8 @@ package com.bookjob.member.service;
 import com.bookjob.common.exception.NotFoundException;
 import com.bookjob.common.exception.UnAuthorizedException;
 import com.bookjob.member.domain.Member;
-import com.bookjob.member.dto.MemberDetailResponse;
-import com.bookjob.member.dto.MyPageResponse;
+import com.bookjob.member.dto.response.MemberDetailResponse;
+import com.bookjob.member.dto.response.MyPageResponse;
 import com.bookjob.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
